requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.25
psycopg[binary,pool]>=3.1.0
apscheduler==3.10.4
google-generativeai>=0.3.0
python-dotenv==1.0.1
httpx==0.26.0
pydantic-settings==2.1.0
psutil>=5.9.0